+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/web/interceptor/CustomLocaleChangeInterceptor.java
@@ -1,6 +1,5 @@
-package com.startingblue.fourtooncookie.global.interceptor;
+package com.startingblue.fourtooncookie.web.interceptor;
 
-import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
