+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/web/interceptor/MemberSignedUpAuthorizationInterceptor.java
@@ -1,4 +1,4 @@
-package com.startingblue.fourtooncookie.global.authorization.member;
+package com.startingblue.fourtooncookie.web.interceptor;
 
 import com.startingblue.fourtooncookie.member.service.MemberService;
 import lombok.extern.slf4j.Slf4j;
